+++ .github/actions/go-test-setup/action.yml
@@ -5,3 +5,7 @@ runs:
       shell: bash
       run: sysctl -w net.core.rmem_max=2500000
       if: ${{ matrix.os == 'ubuntu' }}
+    - name: Run nocover tests. These are tests that require the coverage analysis to be off # See https://github.com/protocol/.github/issues/460
+      shell: bash
+      # This matches only tests with "NoCover" in their test name to avoid running all tests again.
+      run: go test -tags nocover -run NoCover -v ./...
